requirements.txt
bcrypt==3.1.7
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
cycler==0.10.0
fastapi==0.52.0
future==0.18.2
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.12.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
h11==0.9.0
httplib2==0.17.0
httptools==0.1.1
idna==2.9
Jinja2==2.11.1
kiwisolver==1.1.0
livereload==2.6.1
lunr==0.5.6
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mkdocs==1.1
netmiko==3.0.0
nltk==3.4.5
numpy==1.18.2
paramiko==2.7.1
passlib==1.7.2
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.4
PyJWT==1.7.1
PyNaCl==1.3.0
pyparsing==2.4.6
pyserial==3.4
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2019.3
PyYAML==5.3
requests==2.23.0
rsa==4.0
scipy==1.4.1
scp==0.13.2
six==1.14.0
SQLAlchemy==1.3.15
starlette==0.13.2
textfsm==1.1.0
tornado==6.0.4
uritemplate==3.0.1
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
wordcloud==1.6.0
you-get==0.4.1410
